Type correctness in printf, in ktruss(1)'s dump.c.

+++ usr.bin/ktruss/dump.c
@@ -321,7 +321,7 @@ ioctldecode(cmd)
 	*dir = '\0';
 
 	printf(decimal ? ",_IO%s('%c',%ld" : ",_IO%s('%c',%#lx",
-	    dirbuf, (cmd >> 8) & 0xff, cmd & 0xff);
+	    dirbuf, (int)(cmd >> 8) & 0xff, cmd & 0xff);
 	if ((cmd & IOC_VOID) == 0)
 		printf(decimal ? ",%ld)" : ",%#lx)", (cmd >> 16) & 0xff);
 	else
